Move HolidayManagerValueHandler next to the HolidayManager and reduce visibility
